add status bar and messages
